Truncate html by parent tag
Having single root improve consistency.
Removing content outside the root reduce the chance
of unwanted data sneaking into the database.

Parent tag is stored in key with the same name.

Parent tag can be described in 'parent' request parameter,
where any section of the tag is acceptable.

If more than one tag fit the description,
the one with biggest content gets chosen .

If no tag fits the description parent tag gets
chosen as if 'parent' request parameter is empty.

Html is not stripped when 'parent'='None'
or when there are no tags in the html provided.